.FileSel: fixed warning (thanks koldo)
git-svn-id: svn://ultimatepp.org/upp/trunk@2905 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/CtrlLib/FileSel.cpp
@@ -366,7 +366,7 @@ void LazyFileIcons::Do()
 	}
 	int tm = GetTickCount();
 	while(pos < list->GetCount()) {
-		if(GetTickCount() - tm > ptime) {
+		if((int)GetTickCount() - tm > ptime) {
 			ptime = 50;
 			Restart(0);
 			return;
